requirements.txt
flask
neo4j
python-dotenv
markdown
requests
schedule
